#6 makes loading MIMEFileExtensions (v0.1.1) slow:
```console
$ rm ~/.julia/compiled/v1.7/MIMEFileExtensions/*.ji
$ julia -e '@time using MIMEFileExtensions'
  1.088552 seconds (506.06 k allocations: 29.272 MiB, 36.05% compilation time)
$ ls -lh ~/.julia/compiled/v1.7/MIMEFileExtensions
total 208K
-rw-rw-r-- 1 arakaki arakaki 206K Mar 22 21:47 0i4jm_ydM5K.ji
$ julia -e '@time using MIMEFileExtensions'
  0.425975 seconds (478.77 k allocations: 27.534 MiB, 90.62% compilation time)
$ julia -e '@time using MIMEFileExtensions'
  0.425975 seconds (478.77 k allocations: 27.534 MiB, 90.57% compilation time)
```

It looks like the main reason is that `Tuple{Vararg{String}}` is not a terrible type to represent variable-length content.

Using `Vector{String}`:

```console
$ rm ~/.julia/compiled/v1.7/MIMEFileExtensions/*.ji
$ julia -e '@time using MIMEFileExtensions'
  0.443467 seconds (118.20 k allocations: 8.624 MiB, 19.87% compilation time)
$ ls -lh ~/.julia/compiled/v1.7/MIMEFileExtensions
total 208K
-rw-rw-r-- 1 arakaki arakaki 207K Mar 22 21:16 0i4jm_ydM5K.ji
$ julia -e '@time using MIMEFileExtensions'
  0.097309 seconds (90.92 k allocations: 6.949 MiB, 75.95% compilation time)
$ julia -e '@time using MIMEFileExtensions'
  0.097159 seconds (90.91 k allocations: 6.887 MiB, 75.86% compilation time)
```

Using `Vector{Symbol}` (this PR):

```console
$ rm ~/.julia/compiled/v1.7/MIMEFileExtensions/*.ji
$ julia -e '@time using MIMEFileExtensions'
  0.438848 seconds (112.58 k allocations: 8.364 MiB, 18.68% compilation time)
$ ls -lh ~/.julia/compiled/v1.7/MIMEFileExtensions
total 196K
-rw-rw-r-- 1 arakaki arakaki 193K Mar 22 21:15 0i4jm_ydM5K.ji
$ julia -e '@time using MIMEFileExtensions'
  0.091745 seconds (85.30 k allocations: 6.627 MiB, 74.87% compilation time)
$ julia -e '@time using MIMEFileExtensions'
  0.091200 seconds (85.30 k allocations: 6.627 MiB, 75.03% compilation time)
```